requirements.txt
Django==2.0.9
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
PyJWT==1.6.4
pytz==2018.7
